+++ extension/webpack/wrap.config.babel.js
@@ -4,7 +4,7 @@ import baseConfig from './base.config';
 export default baseConfig({
   mode: 'production',
   input: {
-    page: [path.join(__dirname, '../src/browser/extension/inject/pageScript')],
+    page: [path.join(__dirname, '../src/pageScript')],
   },
   output: { path: path.join(__dirname, '../build/tmp') },
   globals: {